Updated to .net 8 and updated nuget packages to newest version

+++ Feedster.DAL/Services/RssFetchService.cs
@@ -240,7 +240,7 @@ namespace Feedster.DAL.Services
                     
                     if (image.Height * image.Width <= highestResolution) continue;
                             
-                    highestResolution = image.Height * image.Width;
+                    highestResolution = (int)(image.Height * image.Width);
                     highestResolutionImage = img;
                 }
                 catch
